third_party/WebKit/LayoutTests/fast/table/display-as-inline.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<p>Test that a table row inside a TABLE that's displayed as inline
    gets wrapped inside an anonymous inline-table.</p>
<span id="first">This</span> text
<table style="display:inline; border-spacing:0;">
    <tr>
        <td style="vertical-align:baseline; padding:0;">should be</td>
    </tr>
</table>
on one <span id="last">line.</span>
<div id="console"></div>
<script>
    var firstTop = document.getElementById('first').getClientRects()[0].top;
    var lastTop = document.getElementById('last').getClientRects()[0].top;
    shouldBe("firstTop", "lastTop");
</script>
